Deduplicate some of the webview code
