+++ src/github/issueOverview.ts
@@ -15,20 +15,9 @@ import { DescriptionNode } from '../view/treeNodes/descriptionNode';
 import { FolderRepositoryManager } from './folderRepositoryManager';
 import { IssueModel } from './issueModel';
 import webviewContent from '../../media/webviewIndex.js';
+import { IRequestMessage, getNonce, WebviewBase } from '../common/webview';
 
-export interface IRequestMessage<T> {
-	req: string;
-	command: string;
-	args: T;
-}
-
-export interface IReplyMessage {
-	seq?: string;
-	err?: any;
-	res?: any;
-}
-
-export class IssueOverviewPanel {
+export class IssueOverviewPanel extends WebviewBase {
 	public static ID: string = 'PullRequestOverviewPanel';
 	/**
 	 * Track the currently panel. Only allow a single panel to exist at a time.
@@ -44,10 +33,6 @@ export class IssueOverviewPanel {
 	protected _item: IssueModel;
 	protected _folderRepositoryManager: FolderRepositoryManager;
 	protected _scrollPosition = { x: 0, y: 0 };
-	private _waitForReady: Promise<void>;
-	private _onIsReady: vscode.EventEmitter<void> = new vscode.EventEmitter();
-
-	protected readonly MESSAGE_UNHANDLED: string = 'message not handled';
 
 	public static async createOrShow(extensionPath: string, folderRepositoryManager: FolderRepositoryManager, issue: IssueModel, descriptionNode: DescriptionNode, toTheSide: Boolean = false) {
 		const activeColumn = toTheSide ?
@@ -75,6 +60,7 @@ export class IssueOverviewPanel {
 	}
 
 	protected constructor(extensionPath: string, column: vscode.ViewColumn, title: string, folderRepositoryManager: FolderRepositoryManager, descriptionNode: DescriptionNode, type: string = IssueOverviewPanel._viewType) {
+		super();
 		this._extensionPath = extensionPath;
 		this._folderRepositoryManager = folderRepositoryManager;
 		this._descriptionNode = descriptionNode;
@@ -91,21 +77,12 @@ export class IssueOverviewPanel {
 			]
 		});
 
-		this._waitForReady = new Promise(resolve => {
-			const disposable = this._onIsReady.event(() => {
-				disposable.dispose();
-				resolve();
-			});
-		});
+		this._webview = this._panel.webview;
+		super.initialize();
 
 		// Listen for when the panel is disposed
 		// This happens when the user closes the panel or when the panel is closed programatically
 		this._panel.onDidDispose(() => this.dispose(), null, this._disposables);
-
-		// Handle messages from the webview
-		this._panel.webview.onDidReceiveMessage(async message => {
-			await this._onDidReceiveMessage(message);
-		}, null, this._disposables);
 
 		this._folderRepositoryManager.onDidChangeActiveIssue(_ => {
 			if (this._folderRepositoryManager && this._item) {
@@ -184,32 +161,12 @@ export class IssueOverviewPanel {
 		return this.updateIssue(issueModel, descriptionNode);
 	}
 
-	protected async _postMessage(message: any) {
-		// Without the following ready check, we can end up in a state where the message handler in the webview
-		// isn't ready for any of the messages we post.
-		await this._waitForReady;
-		this._panel.webview.postMessage({
-			res: message
-		});
-	}
-
-	protected async _replyMessage(originalMessage: IRequestMessage<any>, message: any) {
-		const reply: IReplyMessage = {
-			seq: originalMessage.req,
-			res: message
-		};
-		this._panel.webview.postMessage(reply);
-	}
-
-	protected async _throwError(originalMessage: IRequestMessage<any>, error: any) {
-		const reply: IReplyMessage = {
-			seq: originalMessage.req,
-			err: error
-		};
-		this._panel.webview.postMessage(reply);
-	}
-
 	protected async _onDidReceiveMessage(message: IRequestMessage<any>) {
+		const result = await super._onDidReceiveMessage(message);
+		if (result !== this.MESSAGE_UNHANDLED) {
+			return;
+		}
+
 		switch (message.command) {
 			case 'alert':
 				vscode.window.showErrorMessage(message.args);
@@ -238,9 +195,6 @@ export class IssueOverviewPanel {
 				return this.removeLabel(message);
 			case 'pr.debug':
 				return this.webviewDebug(message);
-			case 'ready':
-				this._onIsReady.fire();
-				return;
 			default:
 				return this.MESSAGE_UNHANDLED;
 		}
@@ -403,13 +357,4 @@ export class IssueOverviewPanel {
 	public getCurrentTitle(): string {
 		return this._panel.title;
 	}
-}
-
-function getNonce() {
-	let text = '';
-	const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-	for (let i = 0; i < 32; i++) {
-		text += possible.charAt(Math.floor(Math.random() * possible.length));
-	}
-	return text;
 }
